backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
asyncpg==0.31.0
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
supabase==2.7.4
